Add half-star ratings and fix duplicate movie slugs

+++ app/Models/Rating.php
@@ -16,7 +16,7 @@ class Rating extends Model
     {
         return [
             'liked' => 'boolean',
-            'stars' => 'integer',
+            'stars' => 'float',
         ];
     }
